backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-anthropic
langchain-community
langchain-pymupdf4llm
docling
googlemaps
httpx
python-dotenv
langsmith
pydantic
python-multipart
